Add a service CA upstream
The use case for this rather hacky addition is that we need to have an
upstream with a CA cert that's common to CoreDNS to test utilizing the
system certs for server cert validation. The primary use case for this
would be if you were running a DNS-over-TLS resolver on the cluster
itself and you wanted to use the service signer.